Refactor bot logic into separate botFunctions.js file
